Move CatchC into a carrier module.
