+++ fused-effects-exceptions.cabal
@@ -19,7 +19,9 @@ tested-with:         GHC == 8.6.4
 library
   hs-source-dirs:      src
   default-language:    Haskell2010
-  exposed-modules:     Control.Effect.Catch
+  exposed-modules:
+    Control.Carrier.Catch
+    Control.Effect.Catch
   build-depends:       base            >= 4.7 && < 5
                      , fused-effects   ^>= 0.5
                      , safe-exceptions >= 0.1 && <1
